Organize header components
refs #31
