+++ stories/HeaderLink.stories.tsx
@@ -1,4 +1,4 @@
-import HeaderLink from "@/components/header-link";
+import HeaderLink from "@/components/header/link";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof HeaderLink> = {
